+++ lib/pagerduty.rb
@@ -22,6 +22,7 @@ require 'pagerduty/models/service'
 require 'pagerduty/models/escalationpolicies'
 require 'pagerduty/models/triggersummarydata'
 require 'pagerduty/models/user'
+require 'pagerduty/models/users'
 require 'pagerduty/models/schedule'
 require 'pagerduty/models/maintenance_window'
 require 'pagerduty/models/notification'
@@ -37,6 +38,15 @@ require 'pagerduty/models/services'
 require 'pagerduty/models/integration'
 
 # Requests
+require 'pagerduty/requests/alerts'
+require 'pagerduty/requests/escalation_policies'
+require 'pagerduty/requests/escalation_rules'
 require 'pagerduty/requests/incident'
+require 'pagerduty/requests/incidents'
 require 'pagerduty/requests/integration'
-#require 'pagerduty/requests/alerts'
+require 'pagerduty/requests/log_entries'
+require 'pagerduty/requests/maintenance_windows'
+require 'pagerduty/requests/notes'
+require 'pagerduty/requests/reports'
+require 'pagerduty/requests/schedules'
+require 'pagerduty/requests/users'
