+++ lib/pagerduty/models/note.rb
@@ -2,7 +2,7 @@ class Note
   include Virtus.model
 
   attribute :id
-  attribute :user, Pagerduty::User
+  attribute :user, Pagerduty::Users::User
   attribute :content
   attribute :created_at
 end
