+++ lib/pagerduty/models/resolvedbyuser.rb
@@ -1,9 +1,12 @@
-class ResolvedByUser
-  include Virtus.model
+class Pagerduty
+  class Users
+    class ResolvedByUser
+      include Virtus.model
 
-  attribute :id
-  attribute :name
-  attribute :email
-  attribute :html_url
+      attribute :id
+      attribute :name
+      attribute :email
+      attribute :html_url
+    end
+  end
 end
-
