+++ lib/pagerduty/models/users.rb
@@ -1,0 +1,11 @@
+class Pagerduty
+  class Users
+    include Virtus.model
+
+    attribute :active_account_users
+    attribute :limit
+    attribute :offset
+    attribute :total
+    attribute :users, Array[Pagerduty::Users::User]
+  end
+end
